Add interface for ProductDataSource to support transactions
