+++ spring-demo/spring-demo-backend/src/main/java/com/vaadin/framework8/samples/backend/repository/ProductRepository.java
@@ -18,6 +18,9 @@ package com.vaadin.framework8.samples.backend.repository;
 import java.util.Collection;
 import java.util.List;
 
+import javax.transaction.Transactional;
+import javax.transaction.Transactional.TxType;
+
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -34,7 +37,7 @@ import com.vaadin.framework8.samples.backend.data.Product;
  * @author Vaadin Ltd
  *
  */
-// @Transactional(TxType.MANDATORY)
+@Transactional(TxType.MANDATORY)
 public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     List<Product> findAllByProductNameContainingIgnoreCaseOrAvailabilityInOrCategoryIn(
